feat: Extract `HangoutDTO`
다른 곳에서도 사용되므로 FetchHangoutsResponseDTO의 내부에 있던 것을 추출합니다.
